Hacky fix to deal with eventual consistency
Cache the last object of a model to be saved, and add it to its
queryset. The cache should time out after 30 seconds, by which time
we assume the database is consistent.

Our cached object might be stale for at worst 30 seconds, if it is
updated immediately elsewhere, but at least the object is present
at all.